services/user_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
passlib[Argon2]
pydantic
pydantic[email]
pydantic-settings
python-dotenv
pytest
pytest-anyio
asyncpg
jsonschema
python-multipart
httpx